Fix #22 by adding intermediate copy step
Fixed by checking if destination folder is on same device and if not,
first copying to a temporary filename, hence preventing truncation if
Eve is killed during a copy (as the file will not be shown as already
downloaded and will be downloaded again on the next run). Deletion of
temp files is done by running cleanup.py as this could potentially take
a while on a large archive, and temp files should be small and rarely
created.