Shitty archive support
eve's code makes some assumptions about threads that are incompatible with archived threads ("threads will be in the threads object to reference", "it is safe to delete a thread from memory if it's not in the list of current threads"). A proper fix would be redesigning those bits and making threads proper objects instead of just 'dicts with some keys that mean stuff i guess idk'.

This code (mostly?) works, so I'll throw it on a side branch so people can use it, but I don't plan to merge it. I just don't really like how I did this.